requirements.txt
pandas==1.3.4
pymongo==4.1.1
python-dotenv==0.21.0
gunicorn==20.1.0
dnspython==2.2.1